tiendasonline v0.1.0 Comprascontroller

Controlador para las funciones para las compras.

Functions

actualizarproducto(cantidad, product)

La función actualizarproducto modificara la cantidad de productos en la tienda. El parametro cantidad sera la cantidad del producto a comprar. El parametro product recibe un mapa del producto.

comprarproducto(tienda, producto, cantidad)

La función comprarproducto realizara la compra del producto. El parametro tienda sera la tienda al a cual pertenece el producto. El parametro producto sera el nombre del producto. El parametro cantidad sera la cantidad del producto a comprar.

Ejemplo: iex>Comprascontroller.comprarproducto(“Tienda1”, “Producto1”, 10)

registrarcompra(product, cantidad)

La función registrarcompra registrara la compra en la base de datos. El parametro product recibe un mapa del producto. El parametro cantidad sera la cantidad del producto a comprar.