tiendasonline v0.1.0 Controller

Controlador para las funciones de las vistas.

Functions

listarproductos(tienda)

La función listarproductos mostrara todos los productos de una tienda registradas en a base de datos. El parametro tienda sera la tienda al a cual pertenece el producto.

listartiendas()

La función listartiendas mostrara todas las tiendas registradas en a base de datos.