tiendasonline v0.1.0 Productocontroller

Controlador para las funciones para productos.

Functions

borrarproducto(tienda, producto)

La función borrarproducto borrara el producto de la base de datos. El parametro tienda sera la tienda al a cual pertenece el producto. El parametro producto sera el nombre del producto.

Ejemplo: iex>Productocontroller.borrarproducto(“Tienda1”, “Producto1”)

registrarproducto(tienda, producto, precio, cantidad)

La función registrarproducto registrara un nuevo producto en la base de datos. El parametro tienda sera la tienda al a cual pertenece el producto. El parametro producto sera el nombre del producto. El parametro precio sera el precio del producto. El parametro cantidad sera la cantidad del producto a ingresar.

Ejemplo:
iex>Productocontroller.registrarproducto("Tienda1", "Producto1", 1000, 10)

verificarproducto(tienda, producto)

La función verificarproducto consultara si el producto se encuentra en la base de datos. El parametro tienda sera la tienda al a cual pertenece el producto. El parametro producto sera el nombre del producto.