tiendasonline v0.1.0 Tiendacontroller

Controlador para las funciones para tiendas.

Functions

borrarproductos(tienda)

La función borrarproductos borrara todos los productos de una tienda. El parametro tienda recibira la tienda la cual se le borraran los productos.

borrartienda(nombre)

La función borrartienda borrara la tienda en la base de datos. El parametro nombre recibira el nombre de la tienda a borrar.

Ejemplo: iex>Tiendacontroller.borrartienda(“Tienda1”)

registrartienda(nombre)

La función registrartienda registrara una nueva tienda en la base de datos. El parametro nombre recibira el nombre de la tienda a ingresar.

Ejemplo: iex>Tiendacontroller.registrartienda(“Tienda1”)

verificartienda(nombre)

La función verificartienda consultara en la base de datos si la tienda se encuentra registrada. El parametro nombre recibira el nombre de la tienda a consultar.